Added back main function since routing seems difficult. Changed port instead.
